requirements.txt
defusedxml==0.5.0
Django==1.10.5
django-ckeditor==5.2.1
django-material==0.12.5
djangorestframework==3.5.4
djangorestframework-jwt==1.9.0
djangorestframework-xml==1.3.0
gunicorn==19.6.0
olefile==0.44
Pillow==4.0.0
pkg-resources==0.0.0
PyJWT==1.4.2
python-slugify==1.2.1
requests==2.13.0
Unidecode==0.4.20
